artsedge/requirements.txt
le-utils>=0.1.9
ricecooker>=0.6.16
lxml>=4.2.1
requests-cache>=0.4.13
